adjust the project structure
