+++ osometweet/__init__.py
@@ -1,1 +1,3 @@
-from osometweet.api import *
+from .api import OsomeTweet
+from .oauth import OAuthHandler, OAuth1a, OAuth2
+from .fields import ObjectExpansions, ObjectFields, ObjectFieldsBase, UserFields, TweetFields, MediaFields, PollFields, PlaceFields
